Move storage tests to integration
