+++ integration/storagetest/maptest.go
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package testharness verifies that storage interfaces behave correctly
-package testharness
+// Package storagetest verifies that storage interfaces behave correctly
+package storagetest
 
 import (
 	"context"
